[SPARK-52539][CONNECT] Introduce session hooks
### What changes were proposed in this pull request?

This PR introduces the concept of Hooks in the Spark connect session. These hooks are used to register code that should run right before an ExecutePlanRequest is sent and possibly modify it. In the future, we might also consider adding more methods to the Hook type that should be called in other places.

### Why are the changes needed?

The motivation for this change is to allow for injecting additional behavior into spark sessions without having to touch the code of the session itself. The Hook interface is abstract and powerful enough to implement complicated behavior while maintaining a clear separation of concerns.

### Does this PR introduce _any_ user-facing change?

In this PR, the session builder has a new method `_registerHook`. The method is private for now so that we can first experiment with the interface.

### How was this patch tested?

I added unit tests that check that a registered hook is created and called as expected.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51233 from niklasmohrin/hooks.

Authored-by: Niklas Mohrin <dev@niklasmohrin.de>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>