[SPARK-52583][SQL] Add an Developer API for stringifying values in UserDefinedType
### What changes were proposed in this pull request?

This PR proposes to add a Developer API for stringifying values in UserDefinedType. When casting a Class X instance that UserDefinedType represents in the Catalyst layer, this API will be called to get the string representation. The default implementation of this API is to call `obj.toString` with the AS-IS behavior.

### Why are the changes needed?

The Class X's toString method doesn't always meet the needs of users, and it's difficult for them to override it in some circumstances. For example, it's a JVM implementation or from other dependent projects.

This stringifyValue API gives a user or developer an extra opportunity to do the override.

### Does this PR introduce _any_ user-facing change?

No, it's a developer API addition and the behavior remains

### How was this patch tested?
New test cases

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #51289 from yaooqinn/SPARK-52583.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>