[SPARK-52547][INFRA][FOLLOW-UP] Set default GIT_BRANCH even for scheduled jobs
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/51245 that sets default GIT_BRANCH even for scheduled jobs.

### Why are the changes needed?

The main PR only fixes the case when the workflow is dispatched with inputs. For scheduled jobs, it does not use default values even so it has to be handled separately.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Manually via bash.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51299 from HyukjinKwon/SPARK-52547-followup2.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>