[SPARK-52642][PYTHON] Use daemonWorkers.get(worker) to avoid unexpected out-of-sync between idleWorkers and daemonWorkers
### What changes were proposed in this pull request?

Uses `daemonWorkers.get(worker)` to avoid unexpected out-of-sync between `idleWorkers` and `daemonWorkers`.

In case of out-of-sync, it will write a warning and stop the out-of-sync worker, then try to use the next idle worker or create a new one.

### Why are the changes needed?

Just in case `idleWorkers` and `daemonWorkers` are not synced to avoid `key not found` error.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

The existing tests should pass.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51345 from ueshin/issues/SPARK-52642/daemonWorkers.

Authored-by: Takuya Ueshin <ueshin@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>