[SPARK-52633][SQL] Deduplicate single `Union` child output before `DeduplicateRelations`
### What changes were proposed in this pull request?
Deduplicate single `Union` child output before `DeduplicateRelations`.

### Why are the changes needed?
Consider the following query pattern:
```
    sql("""CREATE TABLE myTable (
          |    col1 STRING,
          |    col2 STRING,
          |    col3 STRING
          |)
          | USING DELTA;""".stripMargin)
    sql("CREATE VIEW myView as SELECT * FROM myTable")
    sql("""SELECT *
          |FROM (
          |    SELECT col1, col2 FROM myTable
          |    UNION
          |    SELECT col3, col2 FROM myTable
          |    UNION
          |    SELECT col2, col2 FROM myTable
          |) AS subquery;""".stripMargin)
```
When unioning on `myTable`, `col2, col2` from outermost union is going to get deduplicated in `ResolveReferences`, by single child output deduplication algorithm. However, if instead of unioning on `myTable`, we union on `myView`, we will get conflicting `ExprId`s in union branches and will have to wait for `DeduplicateRelations` to kick in. In `DeduplicateRelations`, we are going to place an extra project and alias all output from below in order to resolve `ExprId` conflicts. This will implicitly resolve the issue with single child output conflicts.

However, this algorithm aliases all output whereas the one in `ResolveReferences` aliases only the non-first duplicates. Additionally, `Project` that is added by `DeduplicateRelations` is going to get normalized away in `NormalizePlan` in order to be able to compare fixed-point and single-pass results.

Because of the discrepancy between tables and views as well as potential logical plan mismatches between single-pass and fixed-point, we need to always apply the same algorithm for deduplicating single child output.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added new golden file tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51334 from mihailotim-db/mihailotim-db/union_deduplication.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>