[SPARK-52750][TESTS] Add eventually in flaky test LocalTableScanExec
### What changes were proposed in this pull request?

Add eventually to prevent race conditions with assertion in this test which can lead to flaky failures such as this one https://github.com/alekjarmov/spark/actions/runs/16173233748/job/45655095879#step:10:4203 . This is already done in the other assertions of this test.

### Why are the changes needed?

To not have flaky test failure.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Test only

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #51425 from alekjarmov/flaky-test.

Lead-authored-by: alekjarmov <alek.jarmov@databricks.com>
Co-authored-by: Alek Jarmov <alek.jarmov@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>