[SPARK-52725][CORE] Delay resource profile manager initialization until plugin is loaded
### What changes were proposed in this pull request?
Follow #45052

When we modify the memory configuration in the plugin, the new configuration cannot be updated to the defaultResourceProfile because it is initialized before the plugin is loaded.

For example, the initial spark.executor.memory is set to 8g, but after our plugin processes it, the value is adjusted to 2g. However, the executor's JVM launch parameter (-Xmx) still defaults to 8g according resource profile, which could lead to OOM kills due to onheap memory RSS usage exceeds expectations.

So, the ResourceProfileManager should be initilized after plugin loaded.

### Why are the changes needed?
Make memory configuration changes in plugins take effect at the resource profile.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
ut

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51424 from DenineLu/using-plugin-conf-init-rp.

Authored-by: ludejiu <ludejiu@xiaohongshu.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>