[SPARK-52565][SQL] Enforce ordinal resolution before other sort order expressions
### What changes were proposed in this pull request?
The following query passes, but should fail:
```
SELECT col2 FROM values(1,2) ORDER BY 2,col1
```
As it can be seen in the query, we try to order by 2 which is an ordinal in this case and it should fail with `ORDER_BY_POS_OUT_OF_RANGE`. It doesn't because of Analyzer rule ordering issue.
We first resolve the sort order `col1` using the `ResolveReferencesInSort` rule. It turns out to be a missing input and thus we add it to the `Project` below and add another `Project` on top of sort to retain original output. At that point, plan looks like:
```
Project[col1#x]
+- Sort['UnresolvedOrdinal(2), col1#x]
     +- Project[col2#x, col1#x]
          ...
 ```
Then we go to `ResolveOrdinalInOrderByAndGroupBy` and successfully resolve the `UnresolvedOrdinal` because there are two elements in the `Project` below. This shouldn't happen and we should fail.

In this issue I propose that we fix it by enforcing ordinal resolution before other sort order expressions.

### Why are the changes needed?
To correct behavior of explained query patterns.

### Does this PR introduce _any_ user-facing change?
Yes, query that passed will fail (as it should).

### How was this patch tested?
Added tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #51268 from mihailoale-db/orderbyordinalmissinginputfix.

Authored-by: mihailoale-db <mihailo.aleksic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>