[SPARK-52577][SDP] Add tests for Declarative Pipelines DatasetManager with Hive catalog
### What changes were proposed in this pull request?

Add tests to make sure that the code inside Declarative Pipelines that interacts with the catalog works with Hive in addition to the default catalog.

### Why are the changes needed?

To make sure that the code inside Declarative Pipelines that interacts with the catalog works with Hive in addition to the default catalog.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

It's only tests.

### Was this patch authored or co-authored using generative AI tooling?

Closes #51283 from sryza/hive-tests.

Authored-by: Sandy Ryza <sandy.ryza@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>