[SPARK-52833][SQL] Fix `VariantBuilder.appendFloat`
### What changes were proposed in this pull request?

There is a small bug that `appendFloat` appends 8 bytes, which should be 4 instead. This doesn't cause a correctness issue, because `writeLong` writes the least significant `numBytes` bytes. When int is extended to long, the lower 4 bytes don't change. This may cause an exception when the buffer doesn't have enough capacity.

### Why are the changes needed?

Bug fix.

### Does this PR introduce _any_ user-facing change?

Yes, as stated above.

### How was this patch tested?

Unit test. It would fail (throw an exception) without the fix.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51526 from chenhao-db/fix_float_to_variant.

Authored-by: Chenhao Li <chenhao.li@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>