[SPARK-52860][SQL][TEST] Support V2 write schema evolution in InMemoryTable
### What changes were proposed in this pull request?

Add test for DSV2 Write schema evolution for InMemoryTable, by making it support TableCapability.ACCEPT_ANY_SCHEMA.

### Why are the changes needed?
There is a test on the analysis phase in V2WriteAnalysisSuite to make sure it skips the output resolution if this flag is enabled.  But it is not end-to-end test for this feature (schema evolution on V2Write).  For authors of DSV2 connector, its better to have an end to end test using the demo InMemoryTable to show to how to use this feature.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Ran the new test

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51506 from szehon-ho/unit_test_schema_evolution.

Authored-by: Szehon Ho <szehon.apache@gmail.com>
Signed-off-by: Peter Toth <peter.toth@gmail.com>