[SPARK-52919][SQL] Fix DSv2 Join pushdown to use previously aliased column
### What changes were proposed in this pull request?
There is a bug in join pushdown for DSv2 where we populate the `AttributeMap` with aliased names. If alias is null, we wouldn't populate the map for such attribute and the original column name would be used.

This was wrong, because the column could've been aliased previously, so no new alias is needed in new join. Therefore, instead of using `colName` (which has the information of up to date column name) we had empty map, and would return the original column name (before join pushdown even happened; the column name from scan).

### Why are the changes needed?
Bug in DSv2 join pushdown.

### Does this PR introduce _any_ user-facing change?
No, it is a bug fix.

### How was this patch tested?
Added new test in base suite.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #51622 from PetarVasiljevic-DB/use_previously_aliased_columnName_in_join_pushdown_condition.

Authored-by: Petar Vasiljevic <petar.vasiljevic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>