[SPARK-52709][SQL] Fix parsing of  STRUCT<>
### What changes were proposed in this pull request?
This PR fixes an issue in Spark SQL's parser where empty or nested STRUCT<> types cause incorrect parenthesis tracking and parsing failures. Previously, the parser increased the parenthesis depth counter upon encountering the keyword STRUCT. Due to the operator precedence in the lexer (e.g., NEQ is matched before LT), a construct like STRUCT<> could incorrectly be tokenized as STRUCT and NEQ. This caused the parser to increase the nesting counter without ever decreasing it, eventually resulting in a syntax error.
For example, the following valid query fails under the current logic:
`SELECT cast(null as STRUCT<>), 2 >> 1;`

To fix this, we adjusted the definition of the NEQ token in the SQL lexer so that it no longer matches <> when used in a complex data type. This ensures that the parser correctly interprets the angle brackets as part of a type specification rather than as a comparison operator.

### Why are the changes needed?
bug fix.
By modifying the NEQ token rule to avoid incorrectly matching <> in this context, we ensure that:
-     Empty STRUCT types like STRUCT<> are parsed correctly.
-     Nested and complex STRUCT types are supported without breaking parsing logic.
-     Queries with complex data types and bitwise operations (e.g., >>), are no longer broken due to incorrect token handling.

This change improves Spark SQL’s compatibility with standard SQL syntax and user expectations.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
A new test case has been added to PlanParserSuite.scala to specifically verify that queries containing CAST(null AS STRUCT<>), nested STRUCT<> (if applicable), and the >> operator now parse successfully into the expected logical plan.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes SPARK-52709
Reported by : mihailom-db

Closes #51480 from ManosGEM/SPARK-52709-fix-empty-struct-parsing.

Lead-authored-by: ManosGEM <gemeliarismanolis@gmail.com>
Co-authored-by: Manolis Gemeliaris <gemeliarismanolis@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>