[SPARK-52903][SQL][FOLLOWUP] Fix query failure with LCAs on structs
### What changes were proposed in this pull request?
This PR is a followup to https://github.com/apache/spark/pull/51591. In that PR I proposed that we trim inner aliases before LCA resolution. This PR is improvement on that and trims alias immediately on rule entry.

### Why are the changes needed?
This is necessary in order to fix a long-standing bug in LCA resolution in fixed-point analyzer.

For a query like:
```
SELECT STRUCT(col1 AS a) AS b, b AS c FROM VALUES(1) GROUP BY b
```

LCA resolution never triggers because `eligibleToLiftUp` method always returns false. This is because at the moment of resolution, aggregate node looks like:
```
Aggregate [struct(a, col1#3)], [struct(a, col1#3 AS a#4) AS b#5, lateralAliasReference(b) AS c#6]
```
Because inner aliases are trimmed in grouping expressions but not in aggregate expressions, we never match aggregate expression to a grouping one so the LCA resolution never happens.

### Does this PR introduce _any_ user-facing change?
Query pattern no longer fails with `INTERNAL_ERROR`.

### How was this patch tested?
Added a test case to golden files.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51634 from mihailotim-db/mihailotim-db/fix_lca_bug_fpoint.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>