[SPARK-52944][CORE][SQL][YARN] Fix invalid assertions in tests
### What changes were proposed in this pull request?
This pr fixes some invalid assertions in the test code, mainly addressing two types of issues:

1. Forgetting to use `assert`. For example:

https://github.com/apache/spark/blob/649056fbaff055ff2b2efccd66295bfe2f4ad065/sql/catalyst/src/test/scala/org/apache/spark/sql/types/DataTypeSuite.scala#L342

Here, it is clearly intended to assert that `dt.simpleString equals "struct<c1:string>"`, but the use of `assert` was forgotten, rendering it an invalid expression.

2. Incorrect line breaks in `should be` statements, causing one line of code to be mistakenly treated as two unrelated lines. For example:

https://github.com/apache/spark/blob/649056fbaff055ff2b2efccd66295bfe2f4ad065/core/src/test/scala/org/apache/spark/SortShuffleSuite.scala#L72-L73

Here, it is clearly intended to make an assertion similar to the following:

```
filesCreatedByShuffle.map(_.getName) should be Set("shuffle_0_0_0.data", "shuffle_0_0_0.index")
```

However, due to the incorrect line break, it actually fails to function as an assertion.

Meanwhile, after implementing the aforementioned fixes, this pr also addresses and repairs the failing test cases within it.

### Why are the changes needed?
Fix invalid assertions in tests

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
- Pass Github Actions

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51654 from LuciferYang/unused-expressions.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>