[SPARK-52921][SQL] Specify outputPartitioning for UnionExec for same output partitoning as children operators
### What changes were proposed in this pull request?

This patch updates `outputPartitioning` for `UnionExec` operator for the cases that the output partitionings of its children are the same. So the output partitioning can be known.

### Why are the changes needed?

Currently the output partitioning of `UnionExec` is simply unknown. But if the output partitionings of its children are known to be the same, we can make the union output as the same output partitioning with the children.

But different to the RDD-level `PartitionerAwareUnionRDD`, which only considers the RDD partitioner, SQL operators' `outputPartitioning` doesn't really rely on RDD's partitioner.

Thus, this patch introduces `SQLPartitioningAwareUnionRDD` which is a specified union RDD only for SQL `UnionExec` if the output partitioning is to be the same as its children. Similar to `PartitionerAwareUnionRDD`, it groups the partitions of parent RDDs at corresponding index together but it doesn't require that parent RDDs to have same partitioner.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit test.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #51623 from viirya/union_outputpartition.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>