[SPARK-53003][CORE] Support `strip` in `SparkStringUtils`
### What changes were proposed in this pull request?

This PR aims to support `strip` in `SparkStringUtils`. Note that this new method aims to invoke Scala's `stripPrefix` and `stripSuffix`. So, it serves the existing usage whose removal target is a single-length string.

In addition, this PR adds a new Scalastyle rule to ban `StringUtils.strip(...)` in favor of the built-in implementation.

### Why are the changes needed?

To improve String utility.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51710 from dongjoon-hyun/SPARK-53003.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>