[SPARK-52974][SQL] Don't broadcast dynamicpruning InSubqueries
### What changes were proposed in this pull request?

Currently when AQE is off, DPP `InSubquery`s are planned by the `PlanSubqueries` rule as regular `InSubqueryExec(isDynamicPruning = false)` which causes the results to be broadcasted. But those subquery results are always used at the driver only so there is no need to broadcast the result.

### Why are the changes needed?

Perf improvement when AQE is off + make `PlanDynamicPruningFilters` to be on par with `PlanAdaptiveDynamicPruningFilters`.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs with the revised test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51685 from peter-toth/SPARK-52974-fix-dpp-insubqueries-broadcast.

Authored-by: Peter Toth <peter.toth@gmail.com>
Signed-off-by: Peter Toth <peter.toth@gmail.com>