[SPARK-53072][TESTS] Only run Lock Hardening tests in RocksDBStateStoreLockHardeningSuite
### What changes were proposed in this pull request?

Adding all the necessary methods that we would usually get from RocksDBStateStoreSuite to the Lock Hardening Suite so we no longer have to extend it.

### Why are the changes needed?

Currently, we run all RocksDBStateStoreSuite tests in the Lock Hardening Suite since we extend from that suite, increasing the runtime of the CI. This is undesirable and can significantly increase job runtime.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

It is only a test refactor, as long as the test passes, this is successful.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #51785 from ericm-db/lock-hardening-suite.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>