[SPARK-53170][CORE] Improve `SparkUserAppException` to have `cause` parameter
### What changes were proposed in this pull request?

This PR aims to improve `SparkUserAppException` to have `cause` parameter additionally.

```scala
-private[spark] case class SparkUserAppException(exitCode: Int)
+private[spark] case class SparkUserAppException(exitCode: Int, cause: Throwable = null)
```

### Why are the changes needed?

`SparkUserAppException` is used in many places; `SparkSubmit`, `PythonRunner`, `RRunner`, `SparkPipelines`. Although this is defined as `private[spark]`, a new parameter `cause` can be used to enrich the information from the user application to upper layers internally.

### Does this PR introduce _any_ user-facing change?

No, this is defined as `private[spark]`.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51897 from dongjoon-hyun/SPARK-53170.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>