[SPARK-53155][SQL] Global lower agggregation should not be replaced with a project
This patch fixes the optimization rule `RemoveRedundantAggregates`.

The optimizer rule `RemoveRedundantAggregates` removes redundant lower aggregation from a query plan and replace it with a project of referred non-aggregate expressions. However, if the removed aggregation is a global one, that is not correct because a project is different with a global aggregation in semantics.

For example, if the input relation is empty, a project might be optimized to an empty relation, while a global aggregation will return a single row.

Yes, this fixes a user-facing bug. Previously, a global aggregation under another aggregation might be treated as redundant and replaced as a project with non-aggregation expressions. If the input relation is empty, the replacement is incorrect and might produce incorrect result. This patch adds a new unit test to show the difference.

Unit test, manual test.

No

Closes #51884 from viirya/fix_remove_redundant_agg.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>
(cherry picked from commit 776b68fb03313e1a9d963d66b1e6e46e6707609e)
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>