[MINOR][DOCS] Updated the docstring of DataStreamWriter.foreach() method
Fixed the documentation

### What changes were proposed in this pull request?
There isn't much clarity on the docstring of the DataStreamWriter.foreach() method that takes an object with process() method, as a input. As part of this PR, The docstring is fixed.

### Why are the changes needed?
There isn't much clarity on the docstring of the DataStreamWriter.foreach() method that takes an object with process() method, as a input.

### Does this PR introduce _any_ user-facing change?
Yes. This change modifies the docstring in the DataStreamWriter.foreach() method.

### How was this patch tested?
Simple change on docstring. I haven't tested the changes. Need some support with the testing.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #51316 from nagaarjun-p/master.

Authored-by: Mine4106 <37037296+Mine4106@users.noreply.github.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>