[SPARK-53103][SS] Throw an error if state directory is not empty when query starts
### What changes were proposed in this pull request?

We are implementing a check to ensure the state directory of the checkpoint location is empty before the first batch of a streaming query begins. The new check will throw an error if the directory isn't empty.

### Why are the changes needed?

This will prevent a conflict where background maintenance processes delete files that the new query has just created and prevents sharing state directories.

### Does this PR introduce _any_ user-facing change?

Yes. There is a new error condition ```STATE_STORE_CHECKPOINT_LOCATION_NOT_EMPTY``` that occurs when the state directory is not empty on the first batch of a query.

### How was this patch tested?

Unit tests are added.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #51817 from dylanwong250/SPARK-53103.

Lead-authored-by: Dylan Wong <dylan.wong@databricks.com>
Co-authored-by: dylanwong250 <dylanwong250@gmail.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>