[SPARK-53349][SQL] Optimized XML parser can't handle corrupted files correctly
### What changes were proposed in this pull request?
In https://github.com/apache/spark/pull/51287, we introduced an optimized XML parser, which is more memory-efficient. However, the new parser reads the input stream eagerly on initialization. If the file is corrupted, the error is not caught properly and handled based on the `ignoreCorruptedFiles` option. This PR addresses the issue.

### Why are the changes needed?
Bug fix

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
New tests.

### Was this patch authored or co-authored using generative AI tooling?

Closes #52093 from xiaonanyang-db/SPARK-53349.

Authored-by: Xiaonan Yang <xiaonan.yang@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>