[SPARK-53382][SQL] Fix rCTE bug with malformed recursion
### What changes were proposed in this pull request?

Add special handling in UnionLoopExec of case that prevPlan is null.

### Why are the changes needed?

There is currently a bug where rCTEs crash when the optimizer replaces parts of the plan that return empty results with an empty LocalRelation, which destroys the UnionLoopRef inside the recursion of the UnionLoop, which breaks the assumption in the Exec node that there will be a UnionLoopRef node there. This causes prevPlan to never be defined to a non-null value which causes an spark internal error when tried to evaluate.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New golden file test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52115 from Pajaraja/pavle-martinovic_data/malformedUnionLoopFix.

Authored-by: pavle-martinovic_data <pavle.martinovic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>